# apps/api/v2/swagger/documentation.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v2/selected-calendars": {
      "post": {
        "operationId": "SelectedCalendarsController_addSelectedCalendar",
        "summary": "Enable selected calendar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectedCalendarsInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectedCalendarOutputResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Selected-Calendars"
        ]
      },
      "delete": {
        "operationId": "SelectedCalendarsController_removeSelectedCalendar",
        "summary": "Disable selected calendar",
        "parameters": [
          {
            "name": "integration",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "externalId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "credentialId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectedCalendarOutputResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Selected-Calendars"
        ]
      }
    },
...
    "/v2/slots/reserve": {
      "post": {
        "operationId": "SlotsController_reserveSlot",
        "summary": "Reserve slots for the given time range",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveSlotInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Slots"
        ]
      }
    },
    "/v2/slots/selected-slot": {
      "delete": {
        "operationId": "SlotsController_deleteSelectedSlot",
        "summary": "Unreserve served slots",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "query",
            "description": "UID of the reserved slot",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Slots"
        ]
      }
    },
    "/v2/slots/available": {
      "get": {
        "operationId": "SlotsController_getAvailableSlots",
        "summary": "Get available slots for an event-type",
        "parameters": [
          {
            "name": "startTime",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "endTime",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "eventTypeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Slots"
        ]
      }
    },
...
  },
  "info": {
    "title": "OneHash Cal v2 API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SelectedCalendarsInputDto": {
        "type": "object",
        "properties": {
          "integration": {
            "type": "string"
          },
          "externalId": {
            "type": "string"
          },
          "credentialId": {
            "type": "number"
          }
        },
        "required": [
          "integration",
          "externalId",
          "credentialId"
        ]
      },
      "SelectedCalendarOutputDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number"
          },
          "integration": {
            "type": "string"
          },
          "externalId": {
            "type": "string"
          },
          "credentialId": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "userId",
          "integration",
          "externalId",
          "credentialId"
        ]
      },
      "SelectedCalendarOutputResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/SelectedCalendarOutputDto"
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
...
      "ReserveSlotInput": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}